refactor(sql): remove old compiler
