--- ibis/backends/base/sql/compiler/__init__.py
@@ -1,27 +1,0 @@
-from __future__ import annotations
-
-from ibis.backends.base.sql.compiler.base import DDL, DML
-from ibis.backends.base.sql.compiler.query_builder import (
-    Compiler,
-    Difference,
-    Intersection,
-    Select,
-    SelectBuilder,
-    TableSetFormatter,
-    Union,
-)
-from ibis.backends.base.sql.compiler.translator import ExprTranslator, QueryContext
-
-__all__ = (
-    "Compiler",
-    "Select",
-    "SelectBuilder",
-    "Union",
-    "Intersection",
-    "Difference",
-    "TableSetFormatter",
-    "ExprTranslator",
-    "QueryContext",
-    "DML",
-    "DDL",
-)
