--- ibis/backends/base/sql/registry/__init__.py
@@ -1,34 +1,0 @@
-from __future__ import annotations
-
-from ibis.backends.base.sql.registry.aggregate import reduction
-from ibis.backends.base.sql.registry.helpers import (
-    quote_identifier,
-    sql_type_names,
-    type_to_sql_string,
-)
-from ibis.backends.base.sql.registry.literal import literal, literal_formatters
-from ibis.backends.base.sql.registry.main import (
-    binary_infix_ops,
-    fixed_arity,
-    operation_registry,
-    unary,
-)
-from ibis.backends.base.sql.registry.window import (
-    format_window_frame,
-    time_range_to_range_window,
-)
-
-__all__ = (
-    "quote_identifier",
-    "operation_registry",
-    "binary_infix_ops",
-    "fixed_arity",
-    "literal",
-    "literal_formatters",
-    "sql_type_names",
-    "type_to_sql_string",
-    "reduction",
-    "unary",
-    "format_window_frame",
-    "time_range_to_range_window",
-)
