--- ibis/backends/base/sql/registry/identifiers.py
@@ -1,149 +1,0 @@
-from __future__ import annotations
-
-base_identifiers = [
-    "add",
-    "aggregate",
-    "all",
-    "alter",
-    "and",
-    "api_version",
-    "as",
-    "asc",
-    "avro",
-    "between",
-    "bigint",
-    "binary",
-    "boolean",
-    "by",
-    "cached",
-    "case",
-    "cast",
-    "change",
-    "char",
-    "class",
-    "close_fn",
-    "column",
-    "columns",
-    "comment",
-    "compute",
-    "create",
-    "cross",
-    "data",
-    "database",
-    "databases",
-    "date",
-    "datetime",
-    "decimal",
-    "delimited",
-    "desc",
-    "describe",
-    "distinct",
-    "div",
-    "double",
-    "drop",
-    "else",
-    "end",
-    "escaped",
-    "exists",
-    "explain",
-    "external",
-    "false",
-    "fields",
-    "fileformat",
-    "finalize_fn",
-    "first",
-    "float",
-    "format",
-    "formatted",
-    "from",
-    "full",
-    "function",
-    "functions",
-    "group",
-    "having",
-    "if",
-    "in",
-    "incremental",
-    "init_fn",
-    "inner",
-    "inpath",
-    "insert",
-    "int",
-    "integer",
-    "intermediate",
-    "interval",
-    "into",
-    "invalidate",
-    "is",
-    "join",
-    "last",
-    "left",
-    "like",
-    "limit",
-    "lines",
-    "load",
-    "location",
-    "merge_fn",
-    "metadata",
-    "not",
-    "null",
-    "nulls",
-    "offset",
-    "on",
-    "or",
-    "order",
-    "outer",
-    "overwrite",
-    "parquet",
-    "parquetfile",
-    "partition",
-    "partitioned",
-    "partitions",
-    "prepare_fn",
-    "produced",
-    "rcfile",
-    "real",
-    "refresh",
-    "regexp",
-    "rename",
-    "replace",
-    "returns",
-    "right",
-    "rlike",
-    "row",
-    "schema",
-    "schemas",
-    "select",
-    "semi",
-    "sequencefile",
-    "serdeproperties",
-    "serialize_fn",
-    "set",
-    "show",
-    "smallint",
-    "stats",
-    "stored",
-    "straight_join",
-    "string",
-    "symbol",
-    "table",
-    "tables",
-    "tblproperties",
-    "terminated",
-    "textfile",
-    "then",
-    "timestamp",
-    "tinyint",
-    "to",
-    "true",
-    "uncached",
-    "union",
-    "update_fn",
-    "use",
-    "using",
-    "values",
-    "view",
-    "when",
-    "where",
-    "with",
-]
