+++ ibis/backends/impala/ddl.py
@@ -1,31 +1,305 @@
 from __future__ import annotations
 
-# Copyright 2014 Cloudera Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 import json
 
-from ibis.backends.base.sql.ddl import (
-    AlterTable,
-    BaseDDL,
-    CreateTable,
-    CreateTableWithSchema,
-    DropFunction,
-    format_partition,
-    format_schema,
-    format_tblproperties,
-)
-from ibis.backends.base.sql.registry import type_to_sql_string
+import sqlglot as sg
+
+import ibis.expr.schema as sch
+from ibis.backends.base.sqlglot.datatypes import ImpalaType
+from ibis.backends.base.sqlglot.ddl import DDL, DML, CreateDDL, DropFunction, DropObject
+
+
+class ImpalaBase:
+    dialect = "hive"
+
+    def sanitize_format(self, format):
+        _format_aliases = {"TEXT": "TEXTFILE"}
+
+        if format is None:
+            return None
+        format = format.upper()
+        format = _format_aliases.get(format, format)
+        if format not in ("PARQUET", "AVRO", "TEXTFILE"):
+            raise ValueError(f"Invalid format: {format!r}")
+
+        return format
+
+    def format_dtype(self, dtype):
+        return ImpalaType.to_string(dtype)
+
+    def format_properties(self, props):
+        tokens = []
+        for k, v in sorted(props.items()):
+            tokens.append(f"  '{k}'='{v}'")
+        return "(\n{}\n)".format(",\n".join(tokens))
+
+    def format_tblproperties(self, props):
+        formatted_props = self.format_properties(props)
+        return f"TBLPROPERTIES {formatted_props}"
+
+    def format_serdeproperties(self, props):
+        formatted_props = self.format_properties(props)
+        return f"SERDEPROPERTIES {formatted_props}"
+
+
+class CreateDatabase(ImpalaBase, CreateDDL):
+    def __init__(self, name, path=None, can_exist=False):
+        self.name = name
+        self.path = path
+        self.can_exist = can_exist
+
+    def compile(self):
+        name = self.quote(self.name)
+
+        create_decl = "CREATE DATABASE"
+        create_line = f"{create_decl} {self._if_exists()}{name}"
+        if self.path is not None:
+            create_line += f"\nLOCATION '{self.path}'"
+
+        return create_line
+
+
+class DropDatabase(ImpalaBase, DropObject):
+    _object_type = "DATABASE"
+
+    def __init__(self, name, must_exist=True):
+        super().__init__(must_exist=must_exist)
+        self.name = name
+
+    def _object_name(self):
+        return self.name
+
+
+class CreateTable(ImpalaBase, CreateDDL):
+    def __init__(
+        self,
+        table_name,
+        database=None,
+        external=False,
+        format="parquet",
+        can_exist=False,
+        partition=None,
+        path=None,
+        tbl_properties=None,
+    ):
+        self.table_name = table_name
+        self.database = database
+        self.partition = partition
+        self.path = path
+        self.external = external
+        self.can_exist = can_exist
+        self.format = self.sanitize_format(format)
+        self.tbl_properties = tbl_properties
+
+    @property
+    def _prefix(self):
+        if self.external:
+            return "CREATE EXTERNAL TABLE"
+        else:
+            return "CREATE TABLE"
+
+    def _create_line(self):
+        scoped_name = self.scoped_name(self.table_name, self.database)
+        return f"{self._prefix} {self._if_exists()}{scoped_name}"
+
+    def _location(self):
+        return f"LOCATION '{self.path}'" if self.path else None
+
+    def _storage(self):
+        # By the time we're here, we have a valid format
+        return f"STORED AS {self.format}"
+
+    @property
+    def pieces(self):
+        yield self._create_line()
+        yield from filter(None, self._pieces)
+
+    def compile(self):
+        return "\n".join(self.pieces)
+
+
+class CreateTableWithSchema(CreateTable):
+    def __init__(self, table_name, schema, table_format=None, **kwargs):
+        super().__init__(table_name, **kwargs)
+        self.schema = schema
+        self.table_format = table_format
+
+    @property
+    def _pieces(self):
+        if self.partition is not None:
+            main_schema = self.schema
+            part_schema = self.partition
+            if not isinstance(part_schema, sch.Schema):
+                part_fields = {name: self.schema[name] for name in part_schema}
+                part_schema = sch.Schema(part_fields)
+
+            to_delete = {name for name in self.partition if name in self.schema}
+            fields = {
+                name: dtype
+                for name, dtype in main_schema.items()
+                if name not in to_delete
+            }
+            main_schema = sch.Schema(fields)
+
+            yield self.format_schema(main_schema)
+            yield f"PARTITIONED BY {self.format_schema(part_schema)}"
+        else:
+            yield self.format_schema(self.schema)
+
+        if self.table_format is not None:
+            yield "\n".join(self.table_format.to_ddl())
+        else:
+            yield self._storage()
+
+        yield self._location()
+
+
+class AlterTable(ImpalaBase, DDL):
+    def __init__(
+        self,
+        table,
+        location=None,
+        format=None,
+        tbl_properties=None,
+        serde_properties=None,
+    ):
+        self.table = table
+        self.location = location
+        self.format = self.sanitize_format(format)
+        self.tbl_properties = tbl_properties
+        self.serde_properties = serde_properties
+
+    def _wrap_command(self, cmd):
+        return f"ALTER TABLE {cmd}"
+
+    def _format_properties(self, prefix=""):
+        tokens = []
+
+        if self.location is not None:
+            tokens.append(f"LOCATION '{self.location}'")
+
+        if self.format is not None:
+            tokens.append(f"FILEFORMAT {self.format}")
+
+        if self.tbl_properties is not None:
+            tokens.append(self.format_tblproperties(self.tbl_properties))
+
+        if self.serde_properties is not None:
+            tokens.append(self.format_serdeproperties(self.serde_properties))
+
+        if len(tokens) > 0:
+            return "\n{}{}".format(prefix, "\n".join(tokens))
+        else:
+            return ""
+
+    def compile(self):
+        props = self._format_properties()
+        action = f"{self.table} SET {props}"
+        return self._wrap_command(action)
+
+
+class RenameTable(AlterTable):
+    def __init__(
+        self,
+        old_name: str,
+        new_name: str,
+        old_database: str | None = None,
+        new_database: str | None = None,
+    ):
+        self._old = sg.table(old_name, db=old_database, quoted=True).sql(
+            dialect=self.dialect
+        )
+        self._new = sg.table(new_name, db=new_database, quoted=True).sql(
+            dialect=self.dialect
+        )
+
+    def compile(self):
+        return self._wrap_command(f"{self._old} RENAME TO {self._new}")
+
+
+class DropTable(ImpalaBase, DropObject):
+    _object_type = "TABLE"
+
+    def __init__(self, table_name, database=None, must_exist=True):
+        super().__init__(must_exist=must_exist)
+        self.table_name = table_name
+        self.database = database
+
+    def _object_name(self):
+        return self.scoped_name(self.table_name, self.database)
+
+
+class TruncateTable(ImpalaBase, DDL):
+    _object_type = "TABLE"
+
+    def __init__(self, table_name, database=None):
+        self.table_name = table_name
+        self.database = database
+
+    def compile(self):
+        name = self.scoped_name(self.table_name, self.database)
+        return f"TRUNCATE TABLE {name}"
+
+
+class DropView(DropTable):
+    _object_type = "VIEW"
+
+
+class CTAS(CreateTable):
+    """Create Table As Select."""
+
+    def __init__(
+        self,
+        table_name,
+        select,
+        database=None,
+        external=False,
+        format="parquet",
+        can_exist=False,
+        path=None,
+        partition=None,
+    ):
+        super().__init__(
+            table_name,
+            database=database,
+            external=external,
+            format=format,
+            can_exist=can_exist,
+            path=path,
+            partition=partition,
+        )
+        self.select = select
+
+    @property
+    def _pieces(self):
+        yield self._partitioned_by()
+        yield self._storage()
+        yield self._location()
+        yield "AS"
+        yield self.select
+
+    def _partitioned_by(self):
+        if self.partition is not None:
+            return "PARTITIONED BY ({})".format(
+                ", ".join(self.quote(expr.get_name()) for expr in self.partition)
+            )
+        return None
+
+
+class CreateView(CTAS):
+    """Create a view."""
+
+    def __init__(self, table_name, select, database=None, can_exist=False):
+        super().__init__(table_name, select, database=database, can_exist=can_exist)
+
+    @property
+    def _pieces(self):
+        yield "AS"
+        yield self.select
+
+    @property
+    def _prefix(self):
+        return "CREATE VIEW"
 
 
 class CreateTableParquet(CreateTable):
@@ -57,7 +331,7 @@ class CreateTableParquet(CreateTable):
         elif self.example_table is not None:
             yield f"LIKE {self.example_table}"
         elif self.schema is not None:
-            yield format_schema(self.schema)
+            yield self.format_schema(self.schema)
         else:
             raise NotImplementedError
 
@@ -65,7 +339,7 @@ class CreateTableParquet(CreateTable):
         yield self._location()
 
 
-class DelimitedFormat:
+class DelimitedFormat(ImpalaBase):
     def __init__(
         self,
         path,
@@ -97,10 +371,10 @@ class DelimitedFormat:
 
         if self.na_rep is not None:
             props = {"serialization.null.format": self.na_rep}
-            yield format_tblproperties(props)
+            yield self.format_tblproperties(props)
 
 
-class AvroFormat:
+class AvroFormat(ImpalaBase):
     def __init__(self, path, avro_schema):
         self.path = path
         self.avro_schema = avro_schema
@@ -113,10 +387,10 @@ class AvroFormat:
         schema = "\n".join(x.rstrip() for x in schema.splitlines())
 
         props = {"avro.schema.literal": schema}
-        yield format_tblproperties(props)
+        yield self.format_tblproperties(props)
 
 
-class ParquetFormat:
+class ParquetFormat(ImpalaBase):
     def __init__(self, path):
         self.path = path
 
@@ -158,7 +432,7 @@ class CreateTableAvro(CreateTable):
         yield "\n".join(self.table_format.to_ddl())
 
 
-class LoadData(BaseDDL):
+class LoadData(ImpalaBase, DDL):
     """Generate DDL for LOAD DATA command.
 
     Cannot be cancelled
@@ -186,11 +460,13 @@ class LoadData(BaseDDL):
         overwrite = "OVERWRITE " if self.overwrite else ""
 
         if self.partition is not None:
-            partition = "\n" + format_partition(self.partition, self.partition_schema)
+            partition = "\n" + self.format_partition(
+                self.partition, self.partition_schema
+            )
         else:
             partition = ""
 
-        scoped_name = self._get_scoped_name(self.table_name, self.database)
+        scoped_name = self.scoped_name(self.table_name, self.database)
         return "LOAD DATA INPATH '{}' {}INTO TABLE {}{}".format(
             self.path, overwrite, scoped_name, partition
         )
@@ -218,7 +494,7 @@ class PartitionProperties(AlterTable):
         self.partition_schema = partition_schema
 
     def _compile(self, cmd, property_prefix=""):
-        part = format_partition(self.partition, self.partition_schema)
+        part = self.format_partition(self.partition, self.partition_schema)
         if cmd:
             part = f"{cmd} {part}"
 
@@ -228,6 +504,8 @@ class PartitionProperties(AlterTable):
 
 
 class AddPartition(PartitionProperties):
+    dialect = "hive"
+
     def __init__(self, table, partition, partition_schema, location=None):
         super().__init__(table, partition, partition_schema, location=location)
 
@@ -236,11 +514,15 @@ class AddPartition(PartitionProperties):
 
 
 class AlterPartition(PartitionProperties):
+    dialect = "hive"
+
     def compile(self):
         return self._compile("", "SET ")
 
 
 class DropPartition(PartitionProperties):
+    dialect = "hive"
+
     def __init__(self, table, partition, partition_schema):
         super().__init__(table, partition, partition_schema)
 
@@ -248,18 +530,18 @@ class DropPartition(PartitionProperties):
         return self._compile("DROP")
 
 
-class CacheTable(BaseDDL):
+class CacheTable(ImpalaBase, DDL):
     def __init__(self, table_name, database=None, pool="default"):
         self.table_name = table_name
         self.database = database
         self.pool = pool
 
     def compile(self):
-        scoped_name = self._get_scoped_name(self.table_name, self.database)
+        scoped_name = self.scoped_name(self.table_name, self.database)
         return f"ALTER TABLE {scoped_name} SET CACHED IN '{self.pool}'"
 
 
-class CreateFunction(BaseDDL):
+class CreateFunction(ImpalaBase, DDL):
     _object_type = "FUNCTION"
 
     def __init__(self, func, name=None, database=None):
@@ -268,9 +550,9 @@ class CreateFunction(BaseDDL):
         self.database = database
 
     def _impala_signature(self):
-        scoped_name = self._get_scoped_name(self.name, self.database)
-        input_sig = _impala_input_signature(self.func.inputs)
-        output_sig = type_to_sql_string(self.func.output)
+        scoped_name = self.scoped_name(self.name, self.database)
+        input_sig = ", ".join(map(self.format_dtype, self.func.inputs))
+        output_sig = self.format_dtype(self.func.output)
 
         return f"{scoped_name}({input_sig}) returns {output_sig}"
 
@@ -306,14 +588,14 @@ class CreateUDA(CreateFunction):
         return f"{create_decl} {impala_sig} {joined_tokens}"
 
 
-class DropFunction(DropFunction):
+class DropFunction(ImpalaBase, DropFunction):
     def _impala_signature(self):
-        full_name = self._get_scoped_name(self.name, self.database)
-        input_sig = _impala_input_signature(self.inputs)
+        full_name = self.scoped_name(self.name, self.database)
+        input_sig = ", ".join(map(self.format_dtype, self.inputs))
         return f"{full_name}({input_sig})"
 
 
-class ListFunction(BaseDDL):
+class ListFunction(ImpalaBase, DDL):
     def __init__(self, database, like=None, aggregate=False):
         self.database = database
         self.like = like
@@ -329,6 +611,37 @@ class ListFunction(BaseDDL):
         return statement
 
 
-def _impala_input_signature(inputs):
-    # TODO: varargs '{}...'.format(val)
-    return ", ".join(map(type_to_sql_string, inputs))
+class InsertSelect(ImpalaBase, DML):
+    def __init__(
+        self,
+        table_name,
+        select_expr,
+        database=None,
+        partition=None,
+        partition_schema=None,
+        overwrite=False,
+    ):
+        self.table_name = table_name
+        self.database = database
+        self.select = select_expr
+
+        self.partition = partition
+        self.partition_schema = partition_schema
+
+        self.overwrite = overwrite
+
+    def compile(self):
+        if self.overwrite:
+            cmd = "INSERT OVERWRITE"
+        else:
+            cmd = "INSERT INTO"
+
+        if self.partition is not None:
+            part = self.format_partition(self.partition, self.partition_schema)
+            partition = f" {part} "
+        else:
+            partition = ""
+
+        select_query = self.select
+        scoped_name = self.scoped_name(self.table_name, self.database)
+        return f"{cmd} {scoped_name}{partition}\n{select_query}"
