+++ ibis/backends/impala/tests/snapshots/test_udf/test_create_uda/False/out.sql
@@ -1,4 +1,4 @@
-CREATE AGGREGATE FUNCTION `bar`.`test_name`(string, string) returns bigint location '/foo/bar.so'
+CREATE AGGREGATE FUNCTION `bar`.`test_name`(STRING, STRING) returns BIGINT location '/foo/bar.so'
 init_fn='Init'
 update_fn='Update'
 merge_fn='Merge'
