+++ ibis/backends/impala/tests/snapshots/test_udf/test_create_udf/out.sql
@@ -1,1 +1,1 @@
-CREATE FUNCTION `test_name`(string, string) returns bigint location '/foo/bar.so' symbol='testFunc'
+CREATE FUNCTION `test_name`(STRING, STRING) returns BIGINT location '/foo/bar.so' symbol='testFunc'
