+++ ibis/backends/impala/tests/snapshots/test_udf/test_create_udf_type_conversions/out.sql
@@ -1,1 +1,1 @@
-CREATE FUNCTION `test_name`(string, tinyint, smallint, int) returns bigint location '/foo/bar.so' symbol='testFunc'
+CREATE FUNCTION `test_name`(STRING, TINYINT, SMALLINT, INT) returns BIGINT location '/foo/bar.so' symbol='testFunc'
