+++ ibis/backends/impala/tests/snapshots/test_udf/test_delete_udf_aggregate/out.sql
@@ -1,1 +1,1 @@
-DROP AGGREGATE FUNCTION `test_name`(string, string)
+DROP AGGREGATE FUNCTION `test_name`(STRING, STRING)
