+++ ibis/backends/impala/tests/snapshots/test_udf/test_delete_udf_db/out.sql
@@ -1,1 +1,1 @@
-DROP FUNCTION `test`.`test_name`(string, string)
+DROP FUNCTION `test`.`test_name`(STRING, STRING)
