+++ ibis/backends/impala/tests/snapshots/test_udf/test_delete_udf_if_exists/out.sql
@@ -1,1 +1,1 @@
-DROP FUNCTION IF EXISTS `test_name`(string, string)
+DROP FUNCTION IF EXISTS `test_name`(STRING, STRING)
