+++ ibis/backends/impala/tests/snapshots/test_udf/test_delete_udf_simple/out.sql
@@ -1,1 +1,1 @@
-DROP FUNCTION `test_name`(string, string)
+DROP FUNCTION `test_name`(STRING, STRING)
